Allow custom cowfile name for persistent data
Instead of a fixed name 'cowfile' we allow a custom name which is
predefined in HYBRID_PERSISTENT_FILENAME and prepared to become
overwritten by an XML defintion whose implementation will follow
later. Reason for the change is that a cowfile is visible as
plain data file to the operating system if e.g used on a live
stick. It should be more clear to the user what this file is
good for