Fixed custom args setup for vmdk format
custom args is a hash with option/value pairs. For vmdk the option
could be e.g adapter_type=value and the value is None. This is by
intention because qemu which receives those type of options allows
only the syntax "-o option=value". The kiwi code setting up the
options did not check correctly if there really is a value for
e.g adapter_type