Added pv xen drivers to boot descriptions
Beginning with SLE12 SP2 the default kernel is a pvops
capable kernel. This means it can be used in a Xen HVM full
virtual machine as well as in a PV - paravirtual machine.
However it's required to keep the paravirtual net, block
modules in the initrd. As their location also has move
an update of the boot image descriptions was necessary