Delete %post from kiwi-pxeboot
We do not provide a default config file for the pxe setup anymore.
There is no good default file we can provide here because we don't
know how the user has configured a pxe image. The documentation
should explain how to setup the pxe config and infrastructure with
the help of the static data provided by the kiwi-pxeboot package