Don't prevent an oem image from a custom disk size
An oem disk ist a self expandable disk and thus it normally
doesn't make sense to specify a disk size for it. However there
are filesystems like btrfs which are very hard to precalculate
a minimum required disk space for X bytes of data. In this
situation a user should have the opportunity to specify a
size