Skip local repos if they don't exist
Instead of raising an exception if a local repository does
not exist, we will continue with a warning message. This is
needed for a flawless integration with the buildservice.
Inside of a buildservice environment only those repos are
setup from which packages were taken. Thus it can happen that
a configured repository does not exist in the buildservice
build environment if the resolver would not take any package
from this repository. kiwi should not fail to build such an
image