Delete obsolete support for uboot
arm boot is using grub2 efi images loaded by a firmware. The
firmware could be uboot but due to the non generic way to
setup the board that it loads the firmware all of these tasks
are handled by custom scripts called via the kiwi
editbootconfig / editbootinstall script hooks. Therefore kiwi
itself does not have to setup or install uboot