Delete obsolete GPT fixup code
Related to bnc#825221 code was added to recreate the disk with
a new GPT label. The reason was that parted created a gpt_sync_mbr
partition table when kiwi repartitioned the disk to use the full
geometry. This was an unwanted behavior by parted and causes the
additional code in kiwi. The current parted version behaves
correctly and thus makes this additional code obsolete