Fixed removal of a package in XMLState
If a <package> was unwanted in a <packages> section the former
code sets the name of the package to a python None type. This
causes the rest of the code to operate on potential None types
which causes unexpected TypeError's under certain conditions.
Instead of leaving an inconsistent package section the entire
package section reference should be deleted from the packages
instance holding it. Fixes #76