Fixup cleanup of intermediate config files
kiwi uses e.g etc/hosts from the host system for proper name
resolution during the build. The temporary variant of that
file will be deleted by kiwi at the end of the installation
process. However depending on the package manager and the
distribution it could happen that the intermediate config
file added by kiwi is treated as existing config variant.
In case of rpm a .rpmnew file variant of the config file
is created and that needs to be handled by kiwi. Therefore
this patch adds a private restore method for the .rpmnew
case. It might be needed to add other restore methods to
deal with this issue depending on how other (non rpm) based
package managers handles the situation. Fixes #104