Fixing gce format build
The patch is two fold, first it fixes the name of the raw disk
when it is copied as disk.raw. Second it fixes the content list
of the tarball to be dynamicly build instead of a static allocation.
Reason for this change is that the list of files depends on the
XML description whether or not a gce disk tag is configured.
Fixes #113