Updated xml_state
The changes here modify the method get_users in xml_state in order
to provide a simple list of users instead of nested lists. Also a new
method to get the groups for a given user has been included in
xml_state. This is done in order to slim down and facilitate the
code in kiwi/system/setup.py. This way users and groups can be handled
in more natural way.