Updated get_users and get_user_groups methods
Get_users method has been updated in order to be able to append the
groups attribute when a user is defined in multiple <users> sections in
the description file. Appart from groups, all other user attributes
are ignored if the user was already processed when parsing the description
file.

Get_users_groups method has been also updated so it returns a list of
groups with unique values, keeping the description file order.