Detailed schema validation report
This commit adds the execution of a jing process if the validation
fails, that way the user has more information about what is
actually failing during the validation process. If jing command
is not found, it just produces an info message to warn the user
and kiwi finalizes as expected.