Fixed hybrid setup for large ISO images
The isohybrid tool uses fseek() to locate the bootloader files.
That's unfortunately not 64-bit safe. In case of an ISO image
bigger than 4G this leads to 32-bit offset issues. kiwi can
workaround this problem by putting the bootloader files near
the top of the iso image, which is done by a change in the
sortfile of this commit