Added target_removable attribute
Indicate if the target disk for oem images is deployed to a
removable device e.g a USB stick or not. This only affects
the EFI setup if requested and in the end avoids the creation
of a custom boot menu entry in the firmware of the target
machine on first boot. This is related to bnc#993130