Fixup assembling of mdraid array
when udev discovers an mdraid array it partially starts the array.
That is interfering with the mdadm --assemble call by kiwi which
leads to a busy state and an array in inactive state. Therefore
the method should wait until the raid array really exists no
matter if the assembling is started by udev or kiwi's mdadm call.

In addition if the array got assembled but is incomplete because
devices are missing or the timout is fired, an additional call to
start any array that has been partially assembled is required.

pxeRaidAssemble will throw an exception if after this call no
md device with a size > 0 will show up after a timeout. This
Fixes bnc#1000742