Don't prefix bootpath with snapshot path
We call set-default to make the snapshot subvolume the
default root. In this case, grub2 and the kernel will
always use the correct snapshot. Prefixing it with the
snapshot path will only work, if /.snapshots is accessible
and as long as you don't do a rollback. In rollback case,
only the filesystem default is adjusted, but not grub.cfg.
So grub2 would try to load a kernel from an outdated
snapshot and in workst case, the snapshot is already
deleted and no kernel found anymore.