Clean all repositories after using PackageManager
This commit is to ensure the repositories used in SystemPrepare
are not included in the final image. Also if we guarantee that
the prepared image does not contain the repositories needed to
construct the build-root, there is not need to clean all
repositories in SystemSetup before including those marked with
'imageinclude' in the descruption file.