Fixed quadruple_token tokenizer
Information not provided should end up as a None type in the
list in any case. If an empty string was provided via e.g
--add-repo source,type,, the tokenizer puts in an empty string.
If the information was provided as --add-repo source,type a
None type was used. The code handling the information expected
a None type for not provided information, with an empty string
we end up in validation code which e,g complains that an empty
string can't be converted into an integer base repo priority