Allow unix device names in oem-unattended-id setup
In order to allow a raw device name in oem-unattended-id
the /dev tree has been added to search list. This is useful
if e.g a ramdisk device which is not part of any /dev/disk/...
or /dev/mapper device map should be used as target disk for
the deployment. Thus a setup to stick the deployment to
e.g /dev/ram1 would look like this

<oem-unattended-id>ram1</oem-unattended-id>

This Fixes #221