Added runtime check for boot image reference
If an initrd_system different from kiwi is selected for a
vmx (simple disk) image, it does not make sense to setup a reference
to a kiwi boot image description, because no kiwi boot image will be
built. Despite that it does not hurt it's still an inconsistent
setup. This is related to (bsc#1027610)