Fixup check_target_directory_not_in_shared_cache
The runtime check compared the given target path if it starts
with the cache directory /var/cache/kiwi. This however also
matches for e.g /var/cache/kiwi-foo which would be a valid
target directory. This patch changes the matcher in a way to
really check if the target directory points to the same cache
directory structure.