Fixup default behavior of build command
The build command automatically used an existing root tree
from a former build attempt. However this could cause an
inconsistent image if the former build root was not based on
the same image type setup. Thus it is better to allow this
only if the --allow-existing-root option is specified along
with the build command call