Don't print warning report
The auto generated xml_parse.py uses the python warnings module
The unit tests uses the coverage module in py.test to create
a report. The latest py.test update now also creates a warnings
report which is unwanted because some of the unit tests
intentionally causes the creation of a warning as the expected
result but we don't want to see that in a py.test warnings
report. Therfore this patch switches off the creation of that
warnings report