Fix existing root check, fixes #349
This commit fixes the validation of an existing root directory
for the command 'system build'. System build used to create the root
directory before performing the root existance check, thus the
check was always failing in any case. The root directory is created
inside the RootInit class within the 'create' method.

Fixes #349