Avoid download URL verification inside the buildservice
This patch changes the remote URL validation strategy when running
inside the build service. By design, inside the build service build
environment connections to the outside world are not allowed, thus
any validation attempting to do that will fail. With this patch, when
running inside the build service, KIWI will not try to test if any
download URL is reachable.

Fixes #418