Move to dracut for live ISO images
Instead of the kiwi isoboot layout the live layout has been
changed to follow the rules of the live system support offered
by dracut. The initrd used to boot up the system is now a
dracut generated initrd and can be customized via the 'flags'
attribute in the kiwi XML description. In order to meet the
kiwi live requirements a custom dracut module called kiwi-live
is provided along with the changes in kiwi. This Fixes #470