Fixed validation of isohybrid warnings
The list of warning messages is evaluated line by line
and those not matching the ignore warnings list are treated
as errors. However if an empty line exists it did not match
the ignore warnings list but is also not an error. This
patch makes sure only non empty warning information has
an effect