Adding --ingore-rpm-postscript-errors flag
This commit adds the --ignore-rpm-postscript-errors flag,
which enables to omit a failure when zypper returns 107 exit code.
In former zypper versions it %post install script failures zypper
was not returning any error code, thus this flag allows to install
packages with flaky %post scripts as in previous zypper versions.

Fixes #539