Do not match comments and PIs in XSLT templates
I wanted to add a simple vim modeline to my XML description:

<!--
vim: et:sts=2:sw=2
-->

This made kiwi consume insane amounts of memory during the XSLT
transform step. While this may be a bug in my version of lxml, we do not
transform comments on processing instructions in the conversion
templates, so the easiest solution is not to match them.

Signed-off-by: Michal Marek <MichalMarek1@eaton.com>