Added build of python-kiwi for Debian based distro
This is a first step to also build the kiwi project for
Debian based distros. However so far this only builds
the dracut module packages as those are required for
Debian builds on non Debian based build hosts. In the
long run this package setup should be extended to have
kiwi fully packaged for Debian