Fixed yum vs. yum-deprecated binary lookup
When using the yum package manager it could be either provided
as yum or yum-deprecated binary. Because of this the search
method to find the binary needs to know the context from which
the call is performed. This could be either the host system
or the created image root. This Fixes #624