Update build-test descriptions for Kanku
As part of the buildservice there is now also a CI system called Kanku
which allows for image boot/run tests. The only requirement on the image
level which is missing in our build test descriptions is the setup of
the serial console. This patch updates those image descriptions which
could be tested by kanku to provide a serial console at boot time.

The plan is that our build-test images automatically gets boot tested
by the Kanku CI. This should include boot, console-login, reboot,
console login, shutdown. The required job description to do this
needs to be worked out together with the Kanku team