Fixup setup of root_filesystem_is_multipath
The variable is set to a boolean value or None. If set to
False due to a configuration of oem-multipath-scan this
will cause the dracut multipath module to be omitted.
In any other case the presence of the multipath module
decides whether it will be included into the dracut
generated initrd or not.