Refactor iso helper tools
There are several tools available to build iso images.
genisoimage, mkisofs and also xorriso. The history of
the tools shows them to be used in the range of available
Linux distributions. Thus kiwi contains an abstraction
layer to support them but assumed they are option
compatible. With the xorriso implementation this is no
longer correct and therefore a refactoring of the iso
tools api is required before support for xorriso can
be added. This references Issue #635