Cleanup hybrid call and use of iso tool options
The extra isohybrid call is only needed if the mkisofs
tool category is used. Thus it should be only visible
at the place where the isofs is created and not as an
extra step in the builder tasks. Additionally the handling
of extra options was mkisofs specific and should be
better done as a common meta_data record. The tool
specific options should only appear in the tool specific
class implementations