Use travis stages feature for pypi deployment
Problem ist that the deploy section is called for every
item in the former matrix: setup. This means the pypi
upload was triggered twice for the same archive which
means one target always fails. In order to deploy only
once travis provides a stages feature which is used
in this commit