Fix setup of kiwi_lvm profile variable
kiwi_lvm was always set to true if a volume management system
is in use. However it should only be set to true if the
selected volume management system is lvm. The same applies
to the kiwi_lvmgroup variable which also only makes sense
if the lvm volume management system is used.
This Fixes bsc#1090427