Fixed check for volume group in use
The former implementation evaluates the output of the vgs
command and set the volume group as in use if one of the
listed volume groups on the host contains the group name
set by the image description. This would also match if the
group name set in the image description is e.g 'System' and
the a volume group on the host with name 'SystemVG' exists.
However a conflict only exists on exact match of the name.
The proposed fix is to use the --select feature from vgs
and let it show information on exact match of the vg_name
field. The code in kiwi then just evaluates if the selection
by vgs has a value or not. This Fixes #721